chore: remove Default::default() - Replacewith the type to improve code clarity.
