+++ src/transfers/test_utils.rs
@@ -97,82 +97,3 @@ impl ReplicaSigning for TestReplicaSigning {
         true
     }
 }
-
-/// Produces a genesis balance for a new network.
-#[allow(unused)]
-pub fn get_random_genesis(balance: u64, id: PublicKey) -> Result<CreditAgreementProof> {
-    let threshold = 0;
-    // Nothing comes before genesis, it is a paradox
-    // that it comes from somewhere. In other words, it is
-    // signed over from a "ghost", the keys generated are "ghost" keys,
-    // they come from nothing and can't be verified.
-    // They are unimportant and will be thrown away,
-    // thus the source of random is also unimportant.
-    let mut rng = rand::thread_rng();
-    let bls_secret_key = SecretKeySet::random(threshold, &mut rng);
-    get_genesis(
-        balance,
-        id,
-        bls_secret_key.public_keys(),
-        bls_secret_key.secret_key_share(threshold),
-    )
-}
-
-/// Produces a genesis balance for a new network.
-pub fn get_genesis(
-    balance: u64,
-    id: PublicKey,
-    peer_replicas: PublicKeySet,
-    secret_key_share: SecretKeyShare,
-) -> Result<CreditAgreementProof> {
-    let credit = Credit {
-        id: Default::default(),
-        amount: Token::from_nano(balance),
-        recipient: id,
-        msg: "genesis".to_string(),
-    };
-
-    // actor instances' signatures over > credit <
-
-    let serialised_credit = bincode::serialize(&credit).map_err(Error::Bincode)?;
-
-    let mut credit_sig_shares = BTreeMap::new();
-    let credit_sig_share = secret_key_share.sign(serialised_credit);
-    let _ = credit_sig_shares.insert(0, credit_sig_share);
-
-    println!("Aggregating actor signature..");
-
-    // Combine shares to produce the main signature.
-    let actor_signature = sn_data_types::Signature::Bls(
-        peer_replicas
-            .combine_signatures(&credit_sig_shares)
-            .map_err(|_| Error::CouldNotCombineSignatures)?,
-    );
-
-    let signed_credit = SignedCredit {
-        credit,
-        actor_signature,
-    };
-
-    // replicas signatures over > signed_credit <
-
-    let serialised_credit = bincode::serialize(&signed_credit).map_err(Error::Bincode)?;
-
-    let mut credit_sig_shares = BTreeMap::new();
-    let credit_sig_share = secret_key_share.sign(serialised_credit);
-    let _ = credit_sig_shares.insert(0, credit_sig_share);
-
-    println!("Aggregating replica signature..");
-
-    let debiting_replicas_sig = sn_data_types::Signature::Bls(
-        peer_replicas
-            .combine_signatures(&credit_sig_shares)
-            .map_err(|_| Error::CouldNotCombineSignatures)?,
-    );
-
-    Ok(CreditAgreementProof {
-        signed_credit,
-        debiting_replicas_sig,
-        debiting_replicas_keys: peer_replicas,
-    })
-}
